+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Forge support for Minecraft 1.21.1 and 26.3
+
 ## [0.1.0] - 2026-09-19
 
 ### Added